requirements.txt
transformers
tqdm
# torch
# torchvision
# torchaudio
gradio
einops
timm
# flash_attn
pillow
fastapi
uvicorn
fastapi-jsonrpc
cachetools